+++ tests/test_simulations.py
@@ -53,7 +53,7 @@ class TestSimulation():
         # Check sfield
         sfield = emg3d.get_source_field(
                 self.grid, self.survey.sources['Tx1'].coordinates,
-                frequency=1.0, strength=0)
+                frequency=1.0, strength=1.0)
         assert self.simulation.get_sfield('Tx1', 'f0') == sfield
 
         # Check efield